Added command line option to force unpack. Fixed bug with deleting preload files when reloading.

+++ src/main/java/com/github/otbproject/otbproject/commands/loader/FSCommandLoader.java
@@ -230,6 +230,11 @@ public class FSCommandLoader {
     }
 
     private static boolean move(File source, File dest) {
+        // Check if they're the same
+        if (source.equals(dest)) {
+            return true;
+        }
+        // Make sure overwrites
         if (dest.exists()) {
             dest.delete();
         }
